requirements.txt
anyio==4.9.0
certifi==2025.4.26
exceptiongroup==1.3.0
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
python-telegram-bot==22.0
sniffio==1.3.1
SQLAlchemy==2.0.40
typing_extensions==4.13.2
